+++ CHANGES.md
@@ -1,5 +1,12 @@
 # jirash Changelog
 
+# 1.1.0
+
+- '-a|--assignee' optiotn to 'jirash createissue'
+- `jirash filters`
+- `jirash user USERNAME`
+
+
 # 1.0.0
 
 (first version)